DeathMatch: Put the two lifts in map BPWxBeta into operation.
Thanks to the previously introduced Mover component, this is finally possible
(and easy, too!).

The large main lift is also a very nice example that shows how Movers are
combined with terminals with GUI panels (that in turn are decorated with
PointLight light sources): one for activating the lift at the upper floor,
one for activating at the lower floor, and one for activating the lift from
inside the lift itself.