Build system: Check the proper formatting of the source code within the scope of each build.
The main goal is to avoid common problems such as trailing whitespace, TABs,
improper encoding (non-UTF8), no EOL at end of file, and mixed(!) EOL styles
within a single file.

Files whose newline style does not match the operating system's default are
accounted for statistics, but not counted as errors.

This is still an experimental feature, and disabled by default.