ClipSys: Make all internal constants of `CollisionModelStaticT` external.
These constants were designed for a specific scaling of the world, namely by
factor 25.4 to have the world coordinates in Millimeters.

This step is supposed to help overcoming this single specific scaling:
the goal is to drop the scaling and thus the mm unit, and instead to work with
"world units" everywhere.