General: Remove the `CA3DE_SCALE` constant.
Scaling the map by factor 25.4 was intended to convert inches to millimeters,
so that millimeters would be the base unit everywhere in the map tools and in
the Cafu Engine. This was a stupid idea -- world units should be whatever the
user chooses them to be, and this and the previous commits address the issue.