CaBSP: Separate the methods for creating a BSP tree with and without face splitting.
It turned out that creating a BSP tree *without* face splitting has subtly
different properties than creating a BSP tree *with* face splitting.

Therefore, this commit removes the "minimize face splits" feature from the
creating of the BSP tree; it will be re-introduced in a separate method in the
next commit.