GameSys: Fix human players erroneously blocking Mover moves (make collision models of human players ignore the related entity's orientation).
The collision model (bounding-box) of human players must not rotate (in world-
space) when the player rotates, or else the rotation could accidentally get the
collision model (bounding-box) get stuck in nearby walls.

This is especially bad because rotated bounding-boxes cause "larger" world-
space bounding-boxes returned by `ClipModelT::GetAbsoluteBB()`. This in turn
caused `GetEntityPosCat()` in `CompMover.cpp` to return `POSCAT_IN_SOLID`
possibly only because the player was rotated, making the overall algorithm
(erroneously) think that it was a part's movement that got the player stuck in
solid, blocking the move.