ClipSys: Remove the `TraceRay()` methods from the `ClipModelT` and `ClipWorldT` classes.
This change continues e07d079: After the previous changes, the `TraceRay()`
methods were redundant.
The key idea is to not have one API method for each shape that can be traced,
but to use one of the concrete `TraceSolidT` classes with (eventually) a
single, well-defined method for running a trace.