Font handling: Fix compilation warnings with FreeType 2.5.4 or newer.
Some Linux distributions ship with FreeType 2.5.3 or older, others ship with
FreeType 2.5.4 or newer. As the type of the `rows` and `width` fields of struct
`FT_Bitmap` was changed in version 2.5.4 (see
http://git.savannah.gnu.org/cgit/freetype/freetype2.git/tree/docs/CHANGES for
details), we have to update our code so that it compiles without warnings with
both the newer and older versions of FreeType.