GameSys: Fix client interpolation / approximation of entities with PlayerPhysics components.
Client interpolation / approximation used to be erroneously enabled for *all*
entities that have a PlayerPhysics component since the component's initial
introduction in 1cf0114 of 2013-12-22.

This was principally wrong (interpolation should be enabled on an individual
basis), generally wrong on the server, and especially wrong on clients for the
local human player entity, which is predicted.

This change fixes the problem. The only exception are the 3rd person players on
the clients, for whom interpolation is desirable, but now unavailable. This is
to be addressed at a later time.