Client: Refactor some "repredict" methods.
It was not necessary to pass details of `CaClientWorldT` to
`EngineEntityT::Repredict()` only to work with them there. Instead, we now do
the work directly in the calling method `CaClientWorldT::OurEntity_Repredict()`.

With this change, `EngineEntityT` no longer needs *any* knowledge about
"sequence numbers" (of the game network protocol) and the
`EngineEntityT::Repredict()` method becomes redundant.

This change also removes a vastly outdated text file and adds some `override`
identifiers.