Client: Make interpolation explicit.
Before this change, (NPC) entities on the client were quasi always in their
interpolated state: Only when a new `SC1_FrameInfo` message arrived from the
server were the interpolated values briefly overwritten with those of the
transmitted server frame, only to be immediately restored after the
interpolation target values were updated.

NPC entities being in the interpolated state however interfered with the
local player prediction, because the prediction then erroneously applied player
commands in a world with different values than the server, which in turn caused
prediction mismatches.

With this change, the client world is essentially always in the state of the
last received server frame (with local player prediction applied). Only for
rendering the world are the entities temporarily assigned their interpolated
values.