Client: Refactor `CaClientWorldT::Draw()`.
The previous code was old and awkward, and especially it set up the 3D camera
based on the local human player's state *before* the entities were brought into
the interpolated state for drawing.

At this time, local player entities don't use interpolation yet, but this is
expected to change soon, prepared by this commit.