Client: Add "reconciliation" (the fixing of reprediction mismatches of the local player entity).
This commit finally adds "reconciliation", a feature that has long been missing
from Cafu. It accounts for mismatches after the reprediction of the local human
player entity, "smoothing" the errors out by using interpolation to gradually
correct them.

This is useful when player commands get lost on their way from the client to the
the server, but especially whenever the player is modified on the server
"externally", i.e. by anything that is not directly related to a player command.

The most prominent example for the latter are movers, e.g. moving wings of doors
that push the player, or lifts that the player rides with.

Without reconciliation, such cases come along with heavy visual stuttering,
whereas with this change, these cases are rendered perfectly smooth as if no
networking code was involved at all.