Client: Iterate generically over an entity's interpolators.
The previous approach added one method to the `EntityT` and `ComponentBaseT`
classes for each method that the user code called on the related interpolators,
bloating the respective interfaces.

With this change, we iterate over an entity's components and over a component's
interpolators "generically". This requires the user code to become a bit more
verbose, but decouples the `EntityT` and `ComponentBaseT` interfaces from the
`ApproxBaseT` interface and eliminates the related bloat.