Client: Generalize client effects with new script callback `ComponentBaseT::OnClientFrame()`.
The new script callback makes it possible to modify an entity's state on each
video frame of a client. With this, arbitrary client effects can be added, such
as pulsating light sources, rotating or bopping items, etc.

Note that at this time, when the user's `OnClientFrame()` implementation
modifies some variable in this or another component, e.g. the `Origin`, we rely
on the user having registered the variable also for interpolation beforehand,
typically by having called e.g. `InitClientApprox("Origin")`.

Interpolation is *not* necessary so that setting the custom values works (and
doesn't interfere with it either), but it is only the interpolators that
restore each variable's original ("target") value near the end of
`CaClientWorldT::Draw()`.

This in turn is important so that neither the local human player's prediction
nor the next frame's call to `OnClientFrame()` can experience any rests of the
modified values.