Libs: Add documentation for script callback functions.
This change adds documentation for all script functions that have been logged
by the code introduced in commit 4f28dd6 of 2015-06-21 in a series of test runs.

It also adds documentation for some script functions that have not been logged.
These are functions that are not directly called by C++ code and thus not
covered by the log mechanism, but rather called indirectly (by other script
code).