Map Editor: Add temporary code for creating an initial player prototype prefab.
This change copies the code that the server uses for instantiating new player
entities to the Map Editor, so that newly created maps are instantiated with a
player prototype entity.

This is meant only as a temporary change that saves us the work to manually
reproduce in the Map Editor what the server creates anyways, and will soon be
removed again.