Map Editor: Add special-case handling for the bounding-boxes of children of player prototype entities.
Without this change, the children of player prototype entities (Camera,
FirstPerson, WeaponSound) render in 2D and 3D views in an obtrusive manner that
is bolder than that of the actual player prototype entity.

This is because the natural size of the Camera and WeaponSound bounding-boxes
is relatively large already (+/- 8 in all dimensions). That of the FirstPerson
model is even larger, because the Model component employs a pyramid as a dummy
model whose dimensions are roughly as large as the player's.

A different, more natural solution to this problem would be preferable, e.g.
creating the children of a player prototype only in the `HumanPlayer.lua`
script, but until then, this special-case handling must serve.