Map Editor: Provide the Selection tool with additional reference points for grid-snapping when drag-translating the selected objects.
Besides the bounding-box corners, also the origins of all selected entities are
now available as grid-snapping reference points when drag-translating the
selected objects. (If no entities are in the selection, the bounding-box's
center is used instead.)

With this change it is finally possible to properly translate entities with
children, such as the newly introduced PlayerPrototype prefabs, by their origin.