(Re-)Introduce pre-caching.
We used to use pre-caching in the past, but it went under in the transition to
the entity component system. This change generally restores pre-caching.
It focuses on the "carried weapons" models, which can now be picked up without
affecting frame-rate at all.

More resources can be added for pre-caching at any time.

Closes #139.