Math3D/Brush: Avoid getting stuck in terrain.
The player could occasionally get stuck in terrain, e.g. in map TechDemo.
This change helps, as it is now virtually always possible to get un-stuck again,
e.g. by moving sideways or backwards, but it is still possible to get to a
complete standstill with the forward-key held down.
More work is required to get to the root of this problem.

References #83.