Map Editor: Provide a better user experience if material application with the "Edit Surface Properties" dialog reports problems.
The application of a material to a face or patch is sometimes not possible,
but the problem is possibly identified only late in the `ObtainSurfaceInfo()`
methods. This change provides a new `struct MsgCountsT` that helps to show
wxMessageBox dialogs only once, even if the material was applied to several
faces or patches.