Map Editor: A big revision of all the combinations that are possible when applying a material to a face or patch.
A material can have been picked into the "Edit Surface Properties" dialog in
modes `PlaneProj` or `MatFit`, and it can be applied to faces or patches in
"apply modes" `ApplyNormal`, `ApplyViewAligned`, `ApplyEdgeAligned`, or
`ApplyProjective`, yielding eight combinations for each faces and patches.

This change is the result of a review of all possible combinations.