Map Editor: Use "round near zero" also for other details in `SurfaceInfoT::Save_cmap()`.
The typical range and precision of the values of a `SurfaceInfoT` is limited
and quite well-known. They were all saved with a precision of 6 rather than 9
digits already, but "round near zero" was only applied to the u- and v-vectors.

Other values however benefit from "round near zero" as well, so this change
implements the required functionality.