GameSys: Strictly consider and treat GUIs as parts of other entities.
GUIs are parts of models that are components of other entities.
Just like any other entity (that is not the local human player entity),
their state can only be modified on the server, as prediction only covers
the local human player entity (because only the lhpe's state is saved and
restored during prediction).