Finish the conversion of DokuWiki to reST contents
This change finishes the alignment of the document structure to the old
DokuWiki TOC.

This brings us as close as reasonably possible to the original TOC that was
at https://docs.cafu.de and kept in the now deleted `start.rst` file.

The finished alignment of the original DokuWiki and the now converted
contents marks the completion of the conversion process in the sense that
future changes will go beyond the original contents.

(With the exception of clean-up steps for cases where the converted markup
does not meet the original markup's intend, e.g. in cases where the
alignment of images must be manually revised.)